milestone3/yizhen/ml-workflow/src/data-collector/pyproject.toml
[project]
name = "data-collector"
version = "0.1.0"
description = "DND data collector for MLOps pipeline"
requires-python = ">=3.12"

dependencies = [
    "datasets>=2.19.0",
    "google-cloud-storage>=3.4.1",
    "beautifulsoup4>=4.14.2",
]
